refactor: remove shipments-client.ts wrapper and use API directly
- Remove frontend/lib/shipments-client.ts wrapper file
- Update all imports from @/lib/shipments-client to @/api
- Change all method calls from shipmentsClient to shipmentsApi
- Fix method name collectData -> collectShipmentData
- Add proper type exports in API index
- Update test mocks and imports
- All TypeScript and linting checks pass
